+++ CHANGELOG.md
@@ -67,3 +67,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - 67df33c3: fix empirical bayes assigning values to objects before creating them
 - 62f92b16: fix initial guesses cannot generate starting guesses for parameter with non-finite or nan bounds
 - 88d056ff: fix a bug where undeclared variables caused issue in Empirical Bayes
+- b094aca9: fix inverted SD in the variational bayes method - remove as it is unnecessary